requirements.txt
streamlit>=1.37.0
pdf2image==1.16.3
Pillow>=10.0.0,<11.0.0